# backend/docs/swagger.yaml
basePath: /api/
definitions:
...
  database.Medal:
    properties:
      elimination_id:
        type: integer
      id:
        type: integer
      player_set_id:
        type: integer
      type:
        type: integer
    type: object
...
  endpoint.PutMedalPlayerSetIdById.RequestBody:
    properties:
      player_set_id:
        type: integer
    type: object
...
  response.ErrorIdResponse:
    properties:
      error:
        example: 'invalid ID(1) : sth error'
        type: string
    type: object
  response.ErrorInternalErrorResponse:
    properties:
      error:
        example: 'sth need fix ID(1) : sth error'
        type: string
    type: object
...
host: localhost:80
info:
  contact:
    name: NYCUArchery
    url: https://github.com/NYCUarchery
  description: Gin swagger
  license:
    name: no license yet
  title: Gin swagger
  version: "1.0"
paths:
...
  /medal/{id}:
    get:
      description: Get one medal by id.
      parameters:
      - description: Medal ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success, return one medal
          schema:
            $ref: '#/definitions/database.Medal'
        "400":
          description: invalid medal id
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internal db error / Get Medal By Id
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Show one medal by id.
      tags:
      - Medal
  /medal/elimination/{eliminationid}:
    get:
      description: Get medals of elimination by elimination id.
      parameters:
      - description: elimination ID
        in: path
        name: eliminationid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success, return medals of elimination
          schema:
            items:
              $ref: '#/definitions/database.Medal'
            type: array
        "400":
          description: invalid elimination id
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internal db error / Get Medal Info By Elimination Id
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Show medals of elimination by elimination id.
      tags:
      - Medal
  /medal/playersetid/{id}:
    patch:
      description: Update medal's player set id by id.
      parameters:
      - description: Medal ID
        in: path
        name: id
        required: true
        type: integer
      - description: PlayerSetId
        in: body
        name: PlayerSetId
        required: true
        schema:
          $ref: '#/definitions/endpoint.PutMedalPlayerSetIdById.RequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: success, return nil
        "400":
          description: invalid medal id / invalid playerset id / elimination id of
            medal and playerset is not same
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internal db error / Get Medal By Id / Get Player Set By Id
            / Update Medal PLayer Set Id
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Update medal's player set id by id.
      tags:
      - Medal
...
swagger: "2.0"
...